+++ ui/src/components/expansion-item/QasExpansionItem.vue
@@ -110,6 +110,7 @@ const showContent = ref(false)
 // constants
 const isNestedExpansionItem = inject('isExpansionItem', false)
 const isNestedBox = inject('isBox', false)
+const isDialog = inject('isDialog', false)
 
 // computed
 const hasBadges = computed(() => !!props.badges.length)
@@ -165,10 +166,10 @@ const expansionProps = computed(() => {
 
 const boxProps = computed(() => {
   /**
-   * Caso o QasExpansionItem estiver dentro de um QasBox ou for um QasExpansionItem
+   * Caso o QasExpansionItem estiver dentro de um QasBox, QasDialog ou for um QasExpansionItem
    * dentro de outro QasExpansionItem, o componente terá uma borda.
   */
-  const isBoxed = isNestedBox || isNestedExpansionItem
+  const isBoxed = isNestedBox || isNestedExpansionItem || isDialog
 
   if (!isBoxed) return {}